Only check for statup and shutdown time if checks are enabled in the config.

+++ largescale/run_many_coll.sh
@@ -82,43 +82,46 @@ python3 ../many-collection-test.py ../"$TEST_CFG"
 
 ERROR=$?
 
-kill "$(pgrep mongod)"
+ENABLE_CHECK=$(grep "enable_stats_check" ../"$TEST_CFG" | cut -d = -f 2)
+if [ "$ENABLE_CHECK" == "true" ]; then
+    kill "$(pgrep mongod)"
 
-# If we are using an existing mongod process, we need to start looking from the last
-# "SERVER RESTARTED" message in the logs.
-if [ "$TASK" == "clean-and-populate" ]; then
-    LAST_RESTART=0
-else
-    LAST_RESTART=$(grep -n "SERVER RESTARTED" "$MONGO_LOG"  | tail -1 | cut -d : -f 1)
-fi
+    # If we are using an existing mongod process, we need to start looking from the last
+    # "SERVER RESTARTED" message in the logs.
+    if [ "$TASK" == "clean-and-populate" ]; then
+        LAST_RESTART=0
+    else
+        LAST_RESTART=$(grep -n "SERVER RESTARTED" "$MONGO_LOG"  | tail -1 | cut -d : -f 1)
+    fi
 
-ELAPSED_TIME=0
-# Timeout before exiting the script if WT takes too long to stop.
-TIMEOUT=3600
-echo Waiting for mongod to stop...
-until [ "$ELAPSED_TIME" -ge "$TIMEOUT" ] || tail -n +"$LAST_RESTART" "$MONGO_LOG" | grep "WiredTiger closed" > /dev/null;
-do
-    sleep 1
-    ((ELAPSED_TIME=ELAPSED_TIME+1))
-    echo time elapsed... "$ELAPSED_TIME"s
-done
+    ELAPSED_TIME=0
+    # Timeout before exiting the script if WT takes too long to stop.
+    TIMEOUT=3600
+    echo Waiting for mongod to stop...
+    until [ "$ELAPSED_TIME" -ge "$TIMEOUT" ] || tail -n +"$LAST_RESTART" "$MONGO_LOG" | grep "WiredTiger closed" > /dev/null;
+    do
+        sleep 1
+        ((ELAPSED_TIME=ELAPSED_TIME+1))
+        echo time elapsed... "$ELAPSED_TIME"s
+    done
 
-STARTUP_TIME=$(grep "WiredTiger opened" "$MONGO_LOG" | tail -1 | awk '{print $5}' | grep -Eo '[0-9]{1,}')
-STARTUP_TIME_THRESHOLD=$(grep "max_startup_time" ../"$TEST_CFG" | grep -Eo '[0-9]{1,}')
-if [ "$STARTUP_TIME" -ge "$STARTUP_TIME_THRESHOLD" ]; then
-    echo "Startup time took too long: $STARTUP_TIME ms (max allowed: $STARTUP_TIME_THRESHOLD ms)"
-    ERROR=1
-fi
-
-if [ "$ELAPSED_TIME" -ge "$TIMEOUT" ]; then
-    echo WT took more than "$TIMEOUT"s to shut down. Forcing script to exit...
-    ERROR=1
-else
-    SHUTDOWN_TIME=$(grep "WiredTiger closed" "$MONGO_LOG" | tail -1 | awk '{print $5}' | grep -Eo '[0-9]{1,}')
-    SHUTDOWN_TIME_THRESHOLD=$(grep "max_shutdown_time" ../"$TEST_CFG" | grep -Eo '[0-9]{1,}')
-    if [ "$SHUTDOWN_TIME" -ge "$SHUTDOWN_TIME_THRESHOLD" ]; then
-        echo "Shutdown time took too long: $SHUTDOWN_TIME ms (max allowed: $SHUTDOWN_TIME_THRESHOLD ms)"
+    STARTUP_TIME=$(grep "WiredTiger opened" "$MONGO_LOG" | tail -1 | awk '{print $5}' | grep -Eo '[0-9]{1,}')
+    STARTUP_TIME_THRESHOLD=$(grep "max_startup_time" ../"$TEST_CFG" | grep -Eo '[0-9]{1,}')
+    if [ "$STARTUP_TIME" -ge "$STARTUP_TIME_THRESHOLD" ]; then
+        echo "Startup time took too long: $STARTUP_TIME ms (max allowed: $STARTUP_TIME_THRESHOLD ms)"
         ERROR=1
+    fi
+
+    if [ "$ELAPSED_TIME" -ge "$TIMEOUT" ]; then
+        echo WT took more than "$TIMEOUT"s to shut down. Forcing script to exit...
+        ERROR=1
+    else
+        SHUTDOWN_TIME=$(grep "WiredTiger closed" "$MONGO_LOG" | tail -1 | awk '{print $5}' | grep -Eo '[0-9]{1,}')
+        SHUTDOWN_TIME_THRESHOLD=$(grep "max_shutdown_time" ../"$TEST_CFG" | grep -Eo '[0-9]{1,}')
+        if [ "$SHUTDOWN_TIME" -ge "$SHUTDOWN_TIME_THRESHOLD" ]; then
+            echo "Shutdown time took too long: $SHUTDOWN_TIME ms (max allowed: $SHUTDOWN_TIME_THRESHOLD ms)"
+            ERROR=1
+        fi
     fi
 fi
